[ADD] New module account_chart_update: Update account chart from template:
Adds a wizard to update a company account chart from a template:
  * Generates the new accounts from the template and assigns them to the right company
  * Generates the new taxes and tax codes, changing account assignations
  * Generates the new fiscal positions, changing account and taxes assignations

Before creating the new accounts, taxes, tax codes and fiscal positions, the user can select
which ones must be created.

The user can also choose to update the existing accounts, taxes, tax codes and fiscal positions
from a template.

The problems occurred during the creation/updating of the account chart are shown in the last step.

It is useful when the account law has changed and you want to transfer the new accounts, taxes and
fiscal positions included in the account chart template to an existing company account chart.

Note: Due to the memory limitation of the osv_memory OpenERP objects, only a maximum number of 
objects can be created each time. If a lot of new accounts, taxes, tax codes or fiscal positions
must be created, the wizard should be run several times.