[FIX] project_mrp: correct reference to process.node from sales process
The project_mrp process extends the sales process at
the procurement phase, however that part of the sales
process has now been moved to the `sale_stock` module,
which is not a dependency of `project_mrp` anymore.
There is however little reason to duplicate that
process node, so as a temporary hack, we can copy
its external ID, even though there is no direct
dependency between the two.

For next version we should fix this hack by moving
the node to a common dependency between sale_stock
and project_mrp, such as `sale`.