[REVERT] stock: commit 9557/OPW 600078 assumes that production lot are guaranteed to be set on each relevant move
This assumption is incorrect in OpenERP v7. The WMS
design in OpenERP 7 and previous versions only permits
tracking production lots (aka serial numbers) on
incoming moves, outgoing moves and production moves.
You can specify production lots on internal moves too,
but the system will also generate internal moves for
many reasons, and nothing guarantees that they will
have the right lots assigned.
This is a design decision, and this means that we
must never enforce the availability of specific
production lots.

This features will be available in OpenERP 8 through
the use of an extra concept called "Quants", that
will make this kind of tracking possible.